Merging branch 'Editor im gui' into 'Master'

+++ CodeSolution/Code/AEngine_GameComponents/Particle System/ParticleEmitter.cpp
@@ -27,7 +27,6 @@
 #include "ParticleEmitter.h"
 #include "ParticleSystem.h"
 #include "Utils\AETimeDefs.h"
-#include "Math\AEMathDefs.h"
 
 //Always include last
 #include "Memory\MemLeaks.h"
@@ -51,7 +50,7 @@ ParticleEmitter::~ParticleEmitter()
 void ParticleEmitter::Update(const TimerParams& timerParams, const XMFLOAT3& newPosition)
 {
 	// Work out how much time has passed since the previous update.
-	float elapsedTime = (float)timerParams.m_ElapsedTime;
+	float elapsedTime = timerParams.m_ElapsedTime;
 
 	if (elapsedTime > 0 && m_ParticleSystem != nullptr)
 	{
